+++ pkg/libwebp/enc/quant_enc.c.go
@@ -271,34 +271,34 @@ const MIN_DQ_UV =(-4)
 // We want to emulate jpeg-like behaviour where the expected "good" quality
 // is around q=75. Internally, our "good" middle is around c=50. So we
 // map accordingly using linear piece-wise function
-func QualityToCompression(double c) double {
-  const double linear_c = (c < 0.75) ? c * (2. / 3.) : 2. * c - 1.;
+func QualityToCompression(float64 c) float64 {
+  const float64 linear_c = (c < 0.75) ? c * (2. / 3.) : 2. * c - 1.;
   // The file size roughly scales as pow(quantizer, 3.). Actually, the
   // exponent is somewhere between 2.8 and 3.2, but we're mostly interested
   // in the mid-quant range. So we scale the compressibility inversely to
   // this power-law: quant ~= compression ^ 1/3. This law holds well for
   // low quant. Finer modeling for high-quant would make use of kAcTable[]
   // more explicitly.
-  const double v = pow(linear_c, 1 / 3.);
+  const float64 v = pow(linear_c, 1 / 3.);
   return v;
 }
 
-func QualityToJPEGCompression(double c, double alpha) double {
+func QualityToJPEGCompression(float64 c, float64 alpha) float64 {
   // We map the complexity 'alpha' and quality setting 'c' to a compression
   // exponent empirically matched to the compression curve of libjpeg6b.
   // On average, the WebP output size will be roughly similar to that of a
   // JPEG file compressed with same quality factor.
-  const double amin = 0.30;
-  const double amax = 0.85;
-  const double exp_min = 0.4;
-  const double exp_max = 0.9;
-  const double slope = (exp_min - exp_max) / (amax - amin);
+  const float64 amin = 0.30;
+  const float64 amax = 0.85;
+  const float64 exp_min = 0.4;
+  const float64 exp_max = 0.9;
+  const float64 slope = (exp_min - exp_max) / (amax - amin);
   // Linearly interpolate 'expn' from exp_min to exp_max
   // in the [amin, amax] range.
-  const double expn = (alpha > amax)   ? exp_min
+  const float64 expn = (alpha > amax)   ? exp_min
                       : (alpha < amin) ? exp_max
                                        : exp_max + slope * (alpha - amin);
-  const double v = pow(c, expn);
+  const float64 v = pow(c, expn);
   return v;
 }
 
@@ -347,20 +347,20 @@ func SimplifySegments(/* const */ enc *VP8Encoder) {
 }
 
 // Sets up segment's quantization values, 'base_quant' and filter strengths.
-func VP8SetSegmentParams(/* const */ enc *VP8Encoder, quality float) {
+func VP8SetSegmentParams(/* const */ enc *VP8Encoder, quality float64) {
   var i int
   int dq_uv_ac, dq_uv_dc;
   num_segments := enc.segment_hdr.num_segments;
-  const double amp = SNS_TO_DQ * enc.config.sns_strength / 100. / 128.;
-  const double Q = quality / 100.;
-  const double c_base = enc.config.emulate_jpeg_size
+  const float64 amp = SNS_TO_DQ * enc.config.sns_strength / 100. / 128.;
+  const float64 Q = quality / 100.;
+  const float64 c_base = enc.config.emulate_jpeg_size
                             ? QualityToJPEGCompression(Q, enc.alpha / 255.)
                             : QualityToCompression(Q);
   for i = 0; i < num_segments; i++ {
     // We modulate the base coefficient to accommodate for the quantization
     // susceptibility and allow denser segments to be quantized more.
-    const double expn = 1. - amp * enc.dqm[i].alpha;
-    const double c = pow(c_base, expn);
+    const float64 expn = 1. - amp * enc.dqm[i].alpha;
+    const float64 c = pow(c_base, expn);
     q := (int)(127. * (1. - c));
     assert.Assert(expn > 0.);
     enc.dqm[i].quant = clip(q, 0, 127);
